Extend functionality on inputChoice() and capitalized "i"s
I've extended the functionality of inputChoice by adding a new
inputChoice() method that takes in sets of strings. The old inputchoice
simply called the new method with one string in the set. The first string
will be displayed to the user (usually will by 'y' and 'n') however the
user can type any of the choices in the set and still be recognized.